upgrade to tauri 2
