+++ Cargo.toml
@@ -39,14 +39,14 @@ leptos = { version = "0.6", optional = true }
 
 # only include if not wasm
 [target.'cfg(not(target_family = "wasm"))'.dependencies]
-tauri = { version = "1.6", default-features = false, features = ["wry"] }
+tauri = { version = "^2.2", default-features = false, features = ["wry"] }
 
 [target.'cfg(target_family = "wasm")'.dependencies]
 #tauri-interop-macro = { path = "./tauri-interop-macro", features = ["_wasm"] }
 tauri-interop-macro = { version = "2.1.4", features = [ "_wasm" ] }
 
 [target.'cfg(not(target_family = "wasm"))'.dev-dependencies]
-tauri = "1.6"
+tauri = "^2.2"
 
 [features]
 default = []
