+++ tauri-interop-macro/Cargo.toml
@@ -25,7 +25,7 @@ proc-macro2 = "^1.0"
 proc-macro-error = "1.0.4"
 
 [dev-dependencies]
-tauri = { version = "^1.6", default-features = false, features = ["wry"] }
+tauri = { version = "^2.2", default-features = false, features = ["wry"] }
 log = "^0.4"
 serde = { version = "^1.0", features = [ "derive" ] }
 # required because the intented usage is to use the main crate,
