+++ test-project/api/Cargo.toml
@@ -11,7 +11,7 @@ log = "0.4"
 serde = { version = "1", features = ["derive"] }
 
 [target.'cfg(not(target_family = "wasm"))'.dependencies]
-tauri = { version = "1.6" }
+tauri = { version = "^2.2" }
 
 [target.'cfg(target_family = "wasm")'.dependencies]
 js-sys = { version = "0.3" }
