+++ test-project/src-tauri/.gitignore
@@ -1,4 +1,4 @@
 # Generated by Cargo
 # will have compiled files and executables
 /target/
-
+/gen/
